sas_converter/requirements.txt
pydantic~=2.11.0
python-dotenv~=1.1.0
sqlalchemy~=2.0.41
chardet~=5.2.0
lark~=1.2.2
structlog~=25.1.0
aiofiles~=24.1.0
pytest~=8.3.0
pytest-cov~=6.1.0
# ML / calibration (L2-D ComplexityAgent)
scikit-learn~=1.6.0
numpy~=2.2.0
# LLM boundary resolver
instructor~=1.8.0
tiktoken~=0.9.0
# Azure OpenAI (primary LLM provider — Week 9 migration)
openai~=1.82.0
# Groq (fallback LLM provider — pre-Week 9 primary)
groq~=0.25.0
# RAPTOR semantic clustering (week 5)
sentence-transformers~=4.1.0
torch~=2.7.0
lancedb~=0.22.0
pyarrow~=20.0.0
# Persistence & Index (week 7)
networkx~=3.5.0
duckdb~=1.3.0
pyyaml~=6.0.0
# Robustness (week 9)
psutil~=7.0.0
pytest-asyncio~=0.26.0
# Observability / Azure Monitor (OpenTelemetry-based)
azure-monitor-opentelemetry~=1.6.0
opentelemetry-api~=1.25.0
